pyproject.toml
[tool.poetry]
name = "Bug Tracker"
version = "0.1.0"
description = "Bug tracking system for the whole mitchoulious system :)."
authors = ["Dimitris Argyropoulos <dimitris.argyropoulos@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
sqlalchemy = "^2.0.4"
asyncpg = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyodbc = "^4.0.35"
passlib = "^1.7.4"
pydantic = "^1.10.5"
psycopg2 = "^2.9.5"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
asyncio = "^3.4.3"
pymongo = "^4.3.3"
pyjwt = "^2.6.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
